Refactoring: move all pydantic models in one file
